为了减少不应有的bug,还是硬着头皮改成使用mybatis generator,本次修改发现已知未解决bug:阅读数未能累加;界面问题描述未能正常显示...

+++ src/main/java/com/mralmost/community/dto/QuestionDTO.java
@@ -14,7 +14,7 @@ public class QuestionDTO {
     /**
      * 问题id
      */
-    private Integer id;
+    private Long id;
     /**
      * 问题标题
      */
@@ -48,7 +48,7 @@ public class QuestionDTO {
      */
     private Integer likeCount;
     /**
-     * 问题描述
+     * 问题标签
      */
     private String tag;
     /**
